+++ lib/database/src/so/elements/SoGLModelMatrixElement.c++
@@ -254,7 +254,7 @@ SoGLModelMatrixElement::rotateEltBy(const SbRotation &rotation)
 
     rotation.getValue(axis, angle);
 
-    glRotatef(angle * (180.0 / M_PI), axis[0], axis[1], axis[2]);
+    glRotatef(angle * float(180.0 / M_PI), axis[0], axis[1], axis[2]);
 }
 
 ////////////////////////////////////////////////////////////////////////
